+++ NEWS.md
@@ -5,6 +5,7 @@
 * Bug fix: dbs order was changed to lexicographical order on emr_db.connect.
 * Bug fix: stray logical tracks on non-global dbs were shown.
 * Bug fix: `emr_track.attr.get` returned attribute name instead of value.
+* `emr_track.var.get` now returns `NULL` when variable doesn't exist.
 * Make sure logical tracks cannot be created on non-global db. 
 * Added `db_id` parameter to `emr_track.ls`
 * Removed cpp11 package dependency